% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_countStopCodons.R
\name{countStopCodons}
\alias{countStopCodons}
\title{Screen genomic intervals and count stop codons}
\usage{
countStopCodons(granges, BSgenome, stopCodons = c("TAA", "TAG", "TGA"))
}
\arguments{
\item{granges}{An object (eg: a \code{data.frame}) than can be coerced to a \code{GRanges} object
containing genomic intervals where stop codons will be searched for}

\item{BSgenome}{A \code{BSgenomeObject} specifying the reference genome sequence}

\item{stopCodons}{Vector of sequences defining DNA stop codons (Default: c("TAA", "TAG", "TGA"))}
}
\value{
A \code{data.frame} copy of \code{granges} to which three columns with the number of stop codons
        found in each reading frame are appended.
}
\description{
This function screens genomic interval (eg: intron) for stop codons
}
