% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_countStopCodons.R
\name{detectCodons}
\alias{detectCodons}
\title{Count the number of codons in a sequence}
\usage{
detectCodons(string, whichCodons)
}
\arguments{
\item{string}{A \cite{string} in which a set of codon sequences should be looked for}

\item{whichCodons}{A vector of strings specifying (stop) codons.}
}
\value{
A one-line \code{data.frame} giving the number of stop codons fround in each reading frame.
}
\description{
This function screens genomic interval (eg: intron) for stop codons
}
