% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_DEXSeq-IR.R
\name{DEXSeq4Introns}
\alias{DEXSeq4Introns}
\title{Differential analysis of intronic counts using DEXSeq}
\usage{
DEXSeq4Introns(
  gtf,
  target,
  libType,
  keepAnnotatedRI = TRUE,
  plotDisp = TRUE,
  saveDisp = NULL,
  verbose = FALSE,
  noWarnings = TRUE
)
}
\arguments{
\item{gtf}{Path to a gtf file from which to extract (independent) intron intervals.}

\item{target}{A formatted \code{data.frame} summarising the sequencing experiment (with columns file and condition).}

\item{libType}{Library types of samples from the experiment. Either "single-end" or "paired-end".}

\item{keepAnnotatedRI}{Whether to analyse introns annotated as retained (Default: TRUE).}

\item{plotDisp}{Whether to plot the graph of dispersion on screen (Default: TRUE).}

\item{saveDisp}{A path where to save the plot of dispersions (Default: NULL, no plot will be saved)}

\item{verbose}{Whether to output messages throughout the process (Default: FALSE).}

\item{noWarnings}{Whether to silence possible warnings (Default: TRUE).}
}
\value{
A \code{DEXSeqResults} object as output by function ['DEXSeqResults'] from the DEXSeq package.
}
\description{
This function is a wrapper adapting the DEXSeq (exon) method for differential
analysis to retained introns.
}
\seealso{
['makeTarget']
}
