% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_extractFullLengthReads.R
\name{extractFullLengthReads}
\alias{extractFullLengthReads}
\title{Extract full-length long reads}
\usage{
extractFullLengthReads(
  alignments,
  gtf = NULL,
  cageFile = NULL,
  saveFile = NULL,
  distanceToTSS = 25,
  keepSecondaryAlignment = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{alignments}{Path to long read alignments in a bam file or a \code{GappedReads} object with long read alignments}

\item{gtf}{Path to a reference transcriptome (gtf/gff file) from which to extract annotated TSS}

\item{cageFile}{Path to Transcription Start Sites inferred (eg: from CAGE data)}

\item{saveFile}{Name of the bam file that will be output}

\item{distanceToTSS}{Maximum distance to a TSS (in bp). (Default: 25)}

\item{keepSecondaryAlignment}{Whether secondary alignments should be considered. (Default: FALSE)}

\item{verbose}{Whether some information about read filtering should be displayed.(Default: FALSE)}
}
\value{
Either returns a \code{GappedReads} object if \code{saveFile=NULL} or outputs a bam file
containing read alignments of long reads that may be full-length .
}
\description{
This function extracts long reads whose alignment start position falls nearby
a known Transcription Start Sites (TSS).
For so doing, it makes use of TSS extracted from a given reference transcriptome (annotated TSS)
and/or of TSS inferrred from CAGE data.
At least one gtf file (in \code{gtf} or \code{cageFile}) must be specified.

At the moment, it is intended for direct RNA Nanopore data,
for which molecules are read from the 3' to the 5' end.

NB: it takes into account possible gaps and splicing in read alignments.
}
\details{
Splice-aware genomics alignments are imported into R as a \code{GappedReads} object.
Long reads' alignment start position (most 3' end of the RNA molecule) is identified.
If a TSS from input \code{gtf} or \code{cageFile} files falls within a \code{distanceToTSS}-wide window
around the read start position, then the read is considered to be full-length.
}
\examples{
\dontrun{
# Long read splice-aware alignments on a reference genome (eg: output by Minimap2):
bamFile <- system.file("extdata", "MCF10A_chr2.dont.bam", package = "IRFindeR2")

#Reference transcriptome in a gtf file:
gtf <- system.file("extdata", "Homo_sapiens.GRCh38.97.chr2.gtf", package = "IRFindeR2")

# TSS called from CAGE data carried on a comparable cell line (here MCF7), provided by ENCODE:

# Select full-length reads alignments
# using both annotated transcripts and called TSS (eg: from ENCODE):
full_length_reads <- extractFullLengthReads(bamFile, gtf, cageFile)

# Select full-length reads alignments using only annotated transcripts:
full_length_reads <- extractFullLengthReads(bamFile, gtf)

# Select full-length reads alignments using only called TSS (eg: from ENCODE or in-house data):
full_length_reads <- extractFullLengthReads(bamFile, cageFile)
}

}
