% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_extractIRreads.R
\name{extractIRreads}
\alias{extractIRreads}
\title{Extract reads overlapping at least one intron interval}
\usage{
extractIRreads(
  alignments,
  intronGR = NULL,
  gtf = NULL,
  saveFile = NULL,
  intronMinoverlap = 5,
  overlapType = "any",
  keepSecondaryAlignments = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{alignments}{A path to long read alignments in a bam file or a \code{GappedReads} object with long read alignments.}

\item{intronGR}{A \code{GRanges} object with intronic interval.}

\item{gtf}{Path to a reference transcriptome in a gtf file from which to extract intron intervals.}

\item{saveFile}{Bam file where long reads overlapping an intron will be saved.}

\item{intronMinoverlap}{Minimum overlap (in bp) between a read alignment and an intron. (Default: 5)}

\item{overlapType}{Type of intron overlap to consider.
Either "within" to select only read that contain at least one intron to its full-extent.
Or "any" to consider partial intron overlaps as well. (Default: "any")}

\item{keepSecondaryAlignments}{Whether secondary alignments should be considered.
Only considered when \code{alignments} is a file. (Default: FALSE)}

\item{verbose}{Whether to display some figures along the filtering process. (Default: FALSE)}
}
\value{
Either returns a \code{GappedReads} object if \code{saveFile=NULL} or outputs a bam file
containing input long reads overlapping at least one intron into \code{saveFile} .
}
\description{
This function extracts (long) reads overlapping at least one intron interval.
Intron interval should be either specified as a \code{GRanges} object in \code{intronGR}, or will
be defined internally if a gtf file is provided in \code{gtf}.

NB: it takes into account possible gaps and splicing in read alignments.
}
\seealso{
['createIntronGR']
}
