% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_products.R
\name{fake_products}
\alias{fake_products}
\title{Fake base of products}
\usage{
fake_products(n, seed = 2811)
}
\arguments{
\item{n}{Number of Products to generate}

\item{seed}{the random seed, default is 2811}
}
\value{
a dataframe

A dataframe of fake products.
}
\description{
Fake base of products
}
\examples{
fake_products(10)
}
