% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_transport.R
\name{fake_sondage_answers}
\alias{fake_sondage_answers}
\title{Create fake transport sondage}
\usage{
fake_sondage_answers(
  n = 200,
  x,
  seed = 2811,
  split = FALSE,
  local = c("fr_FR")
)
}
\arguments{
\item{n}{Number of sondage}

\item{x}{Optionnal. fake client data base with "age" column}

\item{seed}{fixe la graine aleatoire}

\item{split}{Logical. Split database in individuals and answers}

\item{local}{the local of the base. Currently supported : "fr_FR" and "en_US".}
}
\description{
Create fake transport sondage
}
\details{
\itemize{
  \item 3 types for each individuals: travail, commerces, loisirs
  \item distance_km. Average distance (km) to target location. Distance is related to age.
  \item transport. Mean of transport to go to target location. Depends on distance.
  \item time_travel_hours. Average duration (hours) to target location. Depends on distance and transport.
}
}
\examples{
answers <- fake_sondage_answers()
\dontrun{
ggplot(answers) + aes(age, log(distance_km), colour = type) + geom_point() +
 geom_smooth() + facet_wrap(~type, scales = "free_y")
}

}
