% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_client.R
\name{fake_support_tickets}
\alias{fake_support_tickets}
\title{Create a fake base of tickets}
\usage{
fake_support_tickets(vol, local = c("en_US", "fr_FR"), seed = 2811)
}
\arguments{
\item{vol}{the number of observations to return}

\item{local}{the local of the base. Currently supported : "fr_FR" and "en_US".}

\item{seed}{the random seed, default is 2811}
}
\description{
A fake base of customer support tickets
}
