% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_products.R
\name{fake_user_feedback}
\alias{fake_user_feedback}
\title{Fake user feedbacks}
\usage{
fake_user_feedback(
  n,
  seed = 2811,
  from = "2012-01-01 00:00:01",
  to = "2020-01-01 00:00:01"
)
}
\arguments{
\item{n}{Number of feedbacks to generate}

\item{seed}{the random seed, default is 2811}

\item{from, to}{the date to cover}
}
\value{
a dataframe
}
\description{
Fake user feedbacks
}
\examples{
fake_user_feedback(10)
}
