% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakethis.R
\name{fake_visits}
\alias{fake_visits}
\title{Title}
\usage{
fake_visits(from = "2017-01-01", to = "2017-12-31", local = c("fr_FR",
  "en_US"), seed = 2811)
}
\arguments{
\item{from, to}{the date to cover}

\item{local}{the local of the base. Currently supported : "fr_FR" and "en_US".}

\item{seed}{the random seed, default is 2811}
}
\description{
Title
}
