% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakir-package.R
\docType{package}
\name{fakir-package}
\alias{fakir-package}
\alias{_PACKAGE}
\alias{fakir}
\title{Create Fake Data in R for tutorials}
\description{
The goal of {fakir} is to provide fake datasets that can be used to teach R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/fakir}
  \item Report bugs at \url{https://github.com/Thinkr-open/fakir/issues}
}

}
\author{
Colin Fay <colin@thinkr.fr>

Sébastien Rochette <sebastien@thinkr.fr>
}
\keyword{internal}
