% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakethis.R
\name{fausse_base_ticket_client}
\alias{fausse_base_ticket_client}
\title{Base ticket client}
\usage{
fausse_base_ticket_client(vol, split = FALSE, seed = 2811)
}
\arguments{
\item{vol}{le nombre de lignes à retourner}

\item{split}{la base doit elle être séparée en deux ?}

\item{seed}{fixe la graine aléatoire}
}
\description{
Une fausse base client de ticket télécom
}
