% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_filterIR.R
\name{filterIRevents}
\alias{filterIRevents}
\title{Call Intron Retention (IR) events captured by NGS data using IRFinder results}
\usage{
filterIRevents(
  irfinderFile,
  minIRratio,
  minSpliceRead = 3,
  minExonToIntronReads = 1,
  minCoverage = 0.5,
  onlyCleanIntrons = TRUE
)
}
\arguments{
\item{irfinderFile}{Path towards a IRFinder-IR-dir.txt file output by IRFinder}

\item{minIRratio}{Minimum retention level (\code{IRratio}) for an intron (value between 0 and 1)}

\item{minSpliceRead}{Minimum number of splice reads (\code{SpliceExact}) supporting the intron (Default value: 3)}

\item{minExonToIntronReads}{Minimum number of reads spanning the exon-to-intron junctions on both side of the intron (Default value: 1)}

\item{minCoverage}{Minimum fraction of the intron that should be covered by reads (Default value: 0.5)}

\item{onlyCleanIntrons}{Whether only introns without any known overlapping event (eg: exon) should be kept (Default: TRUE)}
}
\value{
A \code{data.frame} derived from \code{irfinderFile} with filtered IR events.
}
\description{
This function is intended to call Intron Retention events from IRFinder analysis by applying
advised filters to a IRFinder-IR-dir.txt file.
}
\examples{

\dontrun{

 # Import formatted data.frame from IRfinder output, without any selection:
 filterIRevents(irfinderFile = irfinderFile,
                minIRratio = 0, minSpliceRead = 0,
                minExonToIntronReads = 0, minCoverage = 0)
 # Import formatted data.frame from IRfinder output, without usual intron selection:
 filterIRevents(irfinderFile = irfinderFile)

}
}
