% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_filterIRtranscripts.R
\name{filterIRtranscripts}
\alias{filterIRtranscripts}
\title{Extract assembled Intron-retaining transcripts  using
a reference annotation}
\usage{
filterIRtranscripts(reference.gtf, assembly.gtf, includeAnnotatedIR = TRUE)
}
\arguments{
\item{reference.gtf}{Path to an extensive reference transcript assembly in a gtf file from which introns will be defined}

\item{assembly.gtf}{Path to a (custom) transcript assembly in a gtf file}

\item{includeAnnotatedIR}{\code{Boolean}. Whether to consider introns retained in a reference transcript (Default: TRUE).}
}
\value{
\code{GRanges} or gtf containing intron-retaining transcripts in the assembly
}
\description{
Extract assembled Intron-retaining transcripts  using
a reference annotation
}
