% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annot_findCompatibleTranscripts.R
\name{findCompatibleTranscripts}
\alias{findCompatibleTranscripts}
\title{findCompatibleTranscripts()
Find transcripts compatible with detected intron retention events}
\usage{
findCompatibleTranscripts(
  ir_events,
  gtf,
  transcript_biotypes = NULL,
  outfile = NULL
)
}
\arguments{
\item{ir_events}{Object that can be coerced to a \code{GRanges} object, containing intron genomic intervals}

\item{gtf}{Path to the reference transcriptome where to search for transcripts with intron retention events}

\item{transcript_biotypes}{specifies biotype(s) of annotated transcripts to consider (Default: NULL -ie: all biotypes)}

\item{outfile}{(Defaut: NULL)}
}
\value{
If a valid path is specified in \code{outfile}, saves hypothetical intron-retaining transcripts in gtf format,
        otherwise returns results in a \code{Ranges} object.
}
\description{
This function finds transcripts compatible with detected intron retention events
}
