% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fra_sf.R
\docType{data}
\name{fra_sf}
\alias{fra_sf}
\title{Map of France}
\format{
A data frame with 96 rows, 5 variables and a spatial geometry (MULTIPOLYGON):
\describe{
  \item{OBJECTID}{polygon identifier}
  \item{pays}{country: France}
  \item{region}{region name}
  \item{departement}{departement name}
  \item{id_dpt}{departement id}
  \item{geometry}{polygon geometry}
}
}
\usage{
fra_sf
}
\description{
A map of France as {sf} object. Can be used as dataset or for maps
}
\keyword{datasets}
