% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_web.R
\name{fake_visits}
\alias{fake_visits}
\title{Create a fake support ticket base}
\usage{
fake_visits(
  from = "2017-01-01",
  to = "2017-12-31",
  local = c("en_US", "fr_FR"),
  seed = 2811
)
}
\arguments{
\item{from, to}{the date to cover}

\item{local}{the local of the base. Currently supported : "fr_FR" and "en_US".}

\item{seed}{fixe la graine aleatoire}
}
\value{
A dataframe of fake web visits.
}
\description{
Create a fake support ticket base
}
\examples{
fake_visits()
}
