% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakir-package.R
\docType{package}
\name{fakir-package}
\alias{fakir}
\alias{fakir-package}
\title{fakir: Create Fake Data in R for tutorials}
\description{
Create fake datasets that can be used to teach R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/fakir}
  \item Report bugs at \url{https://github.com/Thinkr-open/fakir/issues}
}

}
\author{
\strong{Maintainer}: Colin Fay \email{colin@thinkr.fr} (0000-0001-7343-1846)

Authors:
\itemize{
  \item Sebastien Rochette \email{sebastien@thinkr.fr} (0000-0002-1565-9313)
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
