% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakir-package.R
\docType{package}
\name{fakir-package}
\alias{fakir}
\alias{fakir-package}
\title{fakir: Generate Fake Datasets for Prototyping and Teaching}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create fake datasets that can be used for prototyping and teaching. This package provides a set of functions to generate fake data for a variety of data types, such as dates, addresses, and names. It can be used for prototyping (notably in 'shiny') or as a tool to teach data manipulation and data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/fakir}
  \item Report bugs at \url{https://github.com/Thinkr-open/fakir/issues}
}

}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})

Authors:
\itemize{
  \item Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
